Move Name out to its own file, support/name.h
