+++ src/wasm.h
@@ -58,39 +58,10 @@
 #include "mixed_arena.h"
 #include "pretty_printing.h"
 #include "support/bits.h"
+#include "support/name.h"
 #include "support/utilities.h"
 
 namespace wasm {
-
-// We use a Name for all of the identifiers. These are IStrings, so they are
-// all interned - comparisons etc are just pointer comparisons, so there is no
-// perf loss. Having names everywhere makes using the AST much nicer (for
-// example, block names are strings and not offsets, which makes composition
-// - adding blocks, removing blocks - easy). One exception is local variables,
-// where we do use indices, as they are a large proportion of the AST,
-// perf matters a lot there, and compositionality is not a problem.
-// TODO: as an optimization, IString values < some threshold could be considered
-//       numerical indices directly.
-
-struct Name : public cashew::IString {
-  Name() : cashew::IString() {}
-  Name(const char* str) : cashew::IString(str, false) {}
-  Name(cashew::IString str) : cashew::IString(str) {}
-  Name(const std::string& str) : cashew::IString(str.c_str(), false) {}
-
-  friend std::ostream& operator<<(std::ostream& o, Name name) {
-    assert(name.str);
-    return o << '$' << name.str; // reference interpreter requires we prefix all names
-  }
-
-  static Name fromInt(size_t i) {
-    return cashew::IString(std::to_string(i).c_str(), false);
-  }
-
-  bool hasSubstring(cashew::IString substring) {
-    return strstr(c_str(), substring.c_str()) != nullptr;
-  }
-};
 
 // An index in a wasm module
 typedef uint32_t Index;
